TST: make each test alone
