+++ xpdconf/tests/test_conf.py
@@ -1,5 +1,4 @@
-import os
-import tempfile
+from databroker.v2 import Broker
 
 
 def test_glbl_dict():
@@ -19,17 +18,8 @@ def test_glbl_dict():
         "_export_tar_dir",
         "archive_base_dir",
         "base_dir",
-        "exp_db",
         "home_dir",
         "blconfig_dir",
     ]:
         assert glbl_dict.get(k)
-
-
-def test_base_dir_from_env():
-    """Test if the base directory can be loaded using the environment variable."""
-    with tempfile.TemporaryDirectory() as d:
-        os.environ["TEST_XPDACQ_BASE_DIR"] = str(d)
-        from xpdconf.conf import glbl_dict
-        assert glbl_dict.get("base_dir") == str(d)
-        os.environ.pop("TEST_XPDACQ_BASE_DIR")
+    assert isinstance(glbl_dict.get("exp_db"), Broker)
